Create ModuleManager, refactor module handling
