+++ src/main/java/com/natem135/hibana/mixin/BlockMixin.java
@@ -1,5 +1,6 @@
 package com.natem135.hibana.mixin;
 
+import com.natem135.hibana.modules.ModuleManager;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
@@ -9,14 +10,13 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import com.natem135.hibana.HibanaClient;
 import com.natem135.hibana.modules.XrayModule;
 
 @Mixin(Block.class)
 public class BlockMixin {
     @Inject(at = { @At("RETURN") }, method = "shouldDrawSide", cancellable = true)
     private static void shouldDrawSide(BlockState state, BlockView world, BlockPos pos, Direction direction, BlockPos blockPos, CallbackInfoReturnable<Boolean> cir) {
-        if(HibanaClient.xrayModule.module_enabled) {
+        if(ModuleManager.xrayModule.module_enabled) {
             cir.setReturnValue(XrayModule.shouldRender(state.getBlock()));
         }
     }
